BUGFIX: conversion from sse to msse was happenning as int, which caused wrapping.
git-svn-id: http://archive-access.svn.sourceforge.net/svnroot/archive-access/trunk/archive-access/projects@1047 69e27eb3-6e27-0410-b9c6-fffd7e226fab

+++ src/java/org/archive/wayback/core/Timestamp.java
@@ -122,7 +122,7 @@ public class Timestamp {
 	 * @param sse
 	 */
 	public void setSse(final int sse) {
-		setDate(new Date(sse * 1000));
+		setDate(new Date(((long)sse) * 1000));
 	}
 	
 	/**
